Started actions and response sections, with other minor updates.
Removed line about /fs/ from the beginning. Added line about the parameters object. Restructured the inital PUT examples paramters from a rename to a move. Made the first pass at the actions and response sections. Added a list of actions, which likely needs a lot of work. Updated GET directory to use the same format as the preceeding examples.