Started overhaul of http request examples, gave example file system, added action links.
Outlined basic file structure to use as exmaple -- needs a graphic.
Broke sentences in the same paragraph out onto their own lines.
Reworded Response section a bit to make more clear the response
format and that read and download are the only exceptions.
Added links to each specific action in the action list.
Refactored the http request and response examples for all the GET
sections to fit the latest convention, and started the first POST
entry. Still need to do the rest, and add Bulkupload section too.